Rename behaviour to handler, moved to handlers package
