+++ src/main/java/io/leonis/torch/component/graph/Gradient.java
