+++ src/main/java/io/leonis/torch/handlers/CycleWindowHandler.java
@@ -1,4 +1,4 @@
-package io.leonis.torch;
+package io.leonis.torch.handlers;
 
 import com.googlecode.lanterna.gui2.*;
 import com.googlecode.lanterna.gui2.TextGUI.Listener;
@@ -7,7 +7,7 @@ import com.googlecode.lanterna.input.*;
 /**
  * @author jeroen.dejong.
  */
-public class CycleWindowBehaviour implements Listener {
+public class CycleWindowHandler implements Listener {
 
   @Override
   public boolean onUnhandledKeyStroke(final TextGUI textGUI, final KeyStroke keyStroke) {
